Format HTML glosses using abbr tag
When outputting to HTML, if a glossing abbreviation formatted
in SmallCaps is recognized, it will be output as HTML's abbr tag
with the gloss displayed upon hover.

Users may specify custom abbreviations as a YAML map in the
document metadata, under the option pandocLingAbbreviations, which
will override any default glossing abbreviations. Abbreviations
must be specified in ALL CAPS. No defaults are currently provided.